requirements.txt
alembic==1.4.3
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.13.0
amqp==5.0.2
aniso8601==8.1.0
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.4
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crcmod==1.7
cryptography==3.2.1
decorator==4.4.2
Flask==1.1.2
Flask-LogConfig==0.4.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
logconfig==0.4.0
logutils==0.3.5
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
Naked==0.1.31
numpy==1.19.4
oss2==2.13.1
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycparser==2.20
pycryptodome==3.9.9
Pygments==2.7.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
PyYAML==5.3.1
qrcode==6.1
redis==3.5.3
requests==2.25.0
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.20
traitlets==5.0.5
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.2.0
